Initial commit
Can search, create, retreive, update and delete entires and live preview changes. You can also edit items but clicking in the them preview (this opens them in the editor).

It prompts for an API key if one hasn't been previously specified and stores it in a cookie, the value is then added to the header of all future AJAX requests.

There is a security issue with the way we are using postMessage() with "*" as an access control value but it's not a problem for now.

This is quite hacky but functional version created during our first integration hack and is our first use of the API.